CoffeeScript source map support.
If you have a file `foo.coffee` or `foo.js.coffee`, when you load `foo.js`, the `X-SourceMap` header is added to the response with value `foo.js.map`.
Inside this source map is a reference to `foo.coffee`. When the browser makes a request to `foo.coffee` it will return the uncompiled CoffeeScript source.
This allows debugging CoffeeScript source in a browser and also works with IntelliJ Remote Javascript debugger too.

Includes tests.

See Issue #39.