Fix a bug of `coffee_engine`
The `coffee_engine` assigns the result of the compiler to `context.data`.
If the `sourceMap` option of coffeescript is disabled,
the result is String, so it doesn't have `js` property.

This means that this engine assigns the `undefined` value to
`context.data` if the `sourceMap` option is disabled.

This commit fixes this bug.